+ improve opus latency (between 200-500 ms of latency) + Extract opus functions into an external javascript (mutli audio support) + git-update-all.sh: check if git lfs is installed and init module for the first checkout + pass desired audio format in init_browser (avoid bugs when mediasource is not supported, and be ready for mutli audio support) + some structure refactoring.
